rename the directory add the autologin truely （may）fix the problem show wrong z-index which does not support backdrop
