+++ frontend/web/README.md
@@ -1,4 +1,4 @@
-# 精弘用户授权中心PC 前端 
+# 精弘用户授权中心 前端 
 
 ## Project setup
 ```
