+++ frontend/web/public/favicon.ico
