+++ frontend/web/src/interface/JsonResponse.ts
