+++ frontend/web/src/interface/backend/app/CheckPre.ts
@@ -1,13 +1,12 @@
 // // path('api/oauth/checkper', oauth.checkper, name='oauth.checkper'),
 import JsonResponse from '@/interface/JsonResponse';
-// interface Request {
-//     studentId: string;
-//     password: string;
-// }
+interface CheckPreRequest {
+    appname: string;
+}
 interface CheckPreResponse extends JsonResponse {
     data: {
         name: string,
     };
-    shortcut: 'ok' | 'pwe' | 'une'  | 'afe';
+    shortcut: 'ok' | 'pwe' | 'une' | 'afe';
 }
-export {CheckPreResponse};
+export { CheckPreResponse, CheckPreRequest };
