+++ frontend/web/src/utils/error.ts
