+++ frontend/web/src/utils/stringFilter.ts
