+++ frontend/web/src/utils/webauthn.ts
