+++ frontend/web/src/views/Activation.vue
