+++ frontend/web/src/views/BaseView.vue
